Add provider dynamic import
